Split out module context and add per-type subclasses
