+++ lib/errors.js
@@ -1,11 +1,10 @@
 module.exports = class ModuleError extends Error {
   constructor(msg, fn = ModuleError, code = fn.name) {
     super(`${code}: ${msg}`)
+
     this.code = code
 
-    if (Error.captureStackTrace) {
-      Error.captureStackTrace(this, fn)
-    }
+    if (Error.captureStackTrace) Error.captureStackTrace(this, fn)
   }
 
   get name() {
