+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bare-module",
-  "version": "6.4.0",
+  "version": "7.0.0-0",
   "description": "Module support for JavaScript",
   "exports": {
     ".": {
@@ -35,7 +35,7 @@
   },
   "homepage": "https://github.com/holepunchto/bare-module#readme",
   "engines": {
-    "bare": ">=1.29.4"
+    "bare": ">=1.30.0"
   },
   "dependencies": {
     "bare-bundle": "^1.3.0",
